Code cleanup
Lots of code cleanup thanks to JetBrains plugins :)
If anything causes issues, it will be this commit as a LOT has changed, everywhere
- Removed multiple redundant type specifications
- Text localised to en-GB & fixed spelling issues (https://raw.githubusercontent.com/wooorm/dictionaries/main/dictionaries/en-GB/index.dic)
- Removed some redundant files
- Multiple merge into patterns (eg: "if (steamIdVal > steamIdMin && steamIdVal < steamIdMax)" to "if (steamIdVal is > steamIdMin and < steamIdMax)"... Never knew this was a thing
- Fixed some possible loss of fraction warnings and tons more.

--> In next commit: Constraints Violations and Redundancies in Code

+++ TcNo-Acc-Switcher-Client/Converters/SteamIdConvert.cs
@@ -78,9 +78,9 @@ namespace TcNo_Acc_Switcher_Client.Converters
             return (int.Parse(input) % 2).ToString();
         }
 
-        private string FloorDivide(string sIn, int divIn)
+        private string FloorDivide(string sIn, double divIn)
         {
-            return Math.Floor((double)(int.Parse(sIn) / divIn)).ToString(CultureInfo.InvariantCulture);
+            return Math.Floor((int.Parse(sIn) / divIn)).ToString(CultureInfo.InvariantCulture);
         }
 
         private string CalcSteamId()
