+++ TcNo-Acc-Switcher-Server/Converters/SteamIdConvert.cs
@@ -80,9 +80,9 @@ namespace TcNo_Acc_Switcher_Server.Converters
             return (int.Parse(input) % 2).ToString();
         }
 
-        private static string FloorDivide(string sIn, int divIn)
+        private static string FloorDivide(string sIn, double divIn)
         {
-            return Math.Floor((double)(int.Parse(sIn) / divIn)).ToString(CultureInfo.InvariantCulture);
+            return Math.Floor((int.Parse(sIn) / divIn)).ToString(CultureInfo.InvariantCulture);
         }
 
         private string CalcSteamId()
